refactor: move logic for getting fee option automatically to useCancelOrder
